requirements.txt
# Data fetching
yfinance>=0.2.40
requests>=2.31

# Data processing
pandas>=2.0
numpy>=1.24
pyarrow>=14.0

# Configuration
pyyaml>=6.0

# UI
streamlit>=1.30
plotly>=5.18

# Testing
pytest>=7.0
